[RA-4567] Read lock functionality reimplemented (#46)
This commit introduces synchronization between the COW file updates and the releasing parent bio.
The idea is to delay reading the COW file until all write operations are completed and tracked in the
COW file. To do that, an additional counter is introduced, which increases when we are ready to release
the parent bio. Sometime after that, the COW file contents update, and the counter decrements. But until
then, all read operations will be suspended.

Note: this commit is for debug purposes