[RA-5413] Debian 11 crashes after the kernel update (#82)
Starting from Debian 11, for some architectures, System.map is available in linux-image-<ver>-dbg
package only. This is a known thing, but it was previously the case for ARM-based distros only - until this
happened on x86. Because of that, no correct System.map was available during the package upgrade,
causing the fallback to /proc/kallsyms and incorrect symbol addresses for the new kernel. This, eventually,
causes the driver to crash after reboot.

As the debug package is absolutely huge, it is suggested to download the deb package without installation
and to extract the System.map only. This PR introduces this fix.