[RA-5394] RHEL7 crashes on unit tests (#83)
There was an issue (elastio#169) when the driver couldn't be unloaded because
something was holding the reference counters. It was found this was caused by
the 'owner' field in the block_device_operations structure. It was then decided
to comment on this field as a temporary solution.

Nevertheless, the problem has returned in the form of a crash on RHEL7 and could also be
reproducible on other distros. The debugging shows these issues are connected, and the
absence of the 'owner' field eventually leads to the kernel crash due to the timing of the
specific events.

When the OS reads the disk through a snapshot device, it acquires and releases control
over it through the open (aka snap_open) and release (aka snap_release) methods of
the block_device_operations. If the driver unloads before the device is released, the
block_device_operations will remain active, and when the OS desires to call the release
method, the kernel will crash.

On the other side, when the snapshot device is destroyed (either due to the destroy ioctl or
transition to the incremental operation mode), the cow thread used to process queued bio
requests is stopped. And if it's stopped while the device is acquired, no bio requests can be
completed, and the OS waits for that with the lock held. As the 'owner' field was commented on,
previously, the driver was able to unload, leaving the block_device_operations active.
Because of that, after some time, the kernel crashed. By uncommenting the 'owner' field back,
we don't allow the driver to unload, but still, bio requests could freeze in the thread queue,
creating a deadlock: processing of the requests has already stopped while the OS is waiting
for them to complete. Because of that, in such a scenario, the reference counter of the driver
was set to 1.

So what needs to be done when the snapshot device is destroyed:

We need to stop the request queue to prevent new bio requests from handling;
We need to wait for all bio requests to be processed to make sure the reference counter is zero;
After that, if any new requests arrive, ignore them to guarantee the same situation from happening.
The 'owner' field should be uncommented back. These changes are introduced in this patch.

Apart from that, an additional wait has been added for the incremental mode transition to
ensure that the driver is synchronized and no pending requests take place. Without this,
from time to time, the driver could return -EBUSY, which was another known issue. This will
also be fixed in this PR.